backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
